feat: add ScenarioGenerator and Scenario class
This adds a ScenarioGenerator class that creates a problem instance, of the class Scenario.
The Scenario class contains operation durations, assignment of operations to machines, and number of operations per job.
A RandomScenarioGenerator class is created to randomly initialize such a Scenario.

When initializing a ScenarioGenerator, max_num_jobs and max_num_ops shall be specified to ensure a static shape for arrays.
When calling the ScenarioGenerator, one can choose an arbitrary number of jobs provided that it is lower than max_num_jobs.