refactor: change configuration to share common code between the 2 JobShop environments
The two environments will be in folders constructive/ and improvement/, with shared methods in the root folder job_shop/ .
The types.py in the root folder is common to both environments, and contains Scenario and a CommonState class, with CommonState containing parameters shared by both environment states.
This allows namely to have, in viewer.py, a visualization that works the same for both environments (the shared parameters used for visualization are those in CommonState).
In constructive/types.py and improvement/types.py, the full State classes are implemented for each environment, originating from CommonState.
Access to the classical job shop environment from other parts of the code is maintained (change of path from 'job_shop' to 'job_shop.constructive' in the files that were using it).