perf: replaced jax.lax.scan by jax.lax.while_loop when computing get_critical_operations
Due to padding, the scan loop was iterating mainly on invalid operations (either sender or receiver_op = -1) which was taking a lot of time.
But due to the structure, we know that all invalid operations are grouped at the end.
With the while loop, we stop at the first invalid operation encountered.
This allows to gain ~30/40% speed on a step.