refactor: cleaner version of get_actions.py and tests, with added documentation
get_actions.py contains the computation of the critical path and important critical block features, as well as the action mask for N5 and N6 neighborhoods.
Documentation in docs/ contains a brief explanation of the problem and the way actions are chosen, in both N5 and N6 neighborhoods. The code refers to it.